requirements.txt
asgiref==3.3.0
Django==3.1.3
pytz==2020.4
sqlparse==0.4.1
Pillow==8.0.1
GitPython==3.1.11
django-progressive-web-app==0.1.1
djangorestframework==3.12.2
